SystemUI: Recents Ram Bar
Add Ram Bar to recents.

Notes.
1) You'll notice the Rambar gets updated frequently in the recents panel
for a few reasons to keep the data as up to date as possible, to update
after swiping away tasks, and also becuase this is now where we pick up
Boolean changes for if the bar is shown and if the all all button is
show. If you think an update is not needed let me know and we can try
to reduce how often we check it.

B) You'll notice the data here does not match the data in
Settings->Apps-> Running. I used the same Math and got values from
the same place as settings but for some reason the data is still different
Mike has me convinced it because the running process here is SystemUI and
it has a larger foot print then Settings. So if you see a mistake again
let me know and we can fix it.

3rd) I'm open for options on theming, at this point both letting the themers
handle it and using a color picker are reasonable ideas. Let the voting begin.

In conclusion I changed kill all to boolean and made it update on the fly
so no systemui kill is needed anymore, removed in my RC commit.

Change-Id: I9e0582b600e1f6a4057d90474dd150db7db5a982

Conflicts:

	core/java/android/provider/Settings.java
	packages/SystemUI/res/layout/system_bar_recent_panel.xml
	packages/SystemUI/res/values/strings.xml
	packages/SystemUI/src/com/android/systemui/recent/RecentsPanelView.java

Signed-off-by: Stevespear426 <stephen.k.spear@gmail.com>